backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-dotenv
requests
gspread
oauth2client
psycopg2-binary
gunicorn
